# Optimize Logs SDK BatchProcessor