@wmdanor 2025-03-28
> @wmdanor, do you think it is ready for review?

It have noticed that CI failed this morning, so I put it back to draft as I did not have enough time to fix it before going to work.

I saw that you have helped me with resolving those issues, thank you for that.

I have addressed your comment below, so it should be ready now.

Other than that, there is race condition comment you left, should I be trying to add some mutex locks for additional security for this in this PR or do we merge this one as good enough?