@pellared 2025-03-28
I think after this is addressed: https://github.com/open-telemetry/opentelemetry-go/pull/6569#discussion_r2019126045 it can be reviewed by others.